Fix pitch parsing (#10)
* update special announcement track to read from local file

* rev version

* load the file properly

* fix sound file loading location

* fix loading local file from docker container

* switching from ADD to COPY

* updating gitignore and adding readme

* updating Dockerfile to target known file during copy

* fixing channel listener

* fixing flux errors

* fixing pitch and speaking rate parsing

+++ src/main/java/com/bob/jr/commands/VoiceCommands.java
@@ -86,7 +86,7 @@ public class VoiceCommands {
                             break;
                         default:
                             if (context.matches(RegexHelper.POS_NEG_DECIMAL_REGEX)) {
-                                final double doubleValue = Integer.parseInt(context) > 0 ?
+                                final double doubleValue = Double.parseDouble(context) > 0 ?
                                         Math.min(Double.parseDouble(context), LimitsHelper.PitchLimits.upperLimit) :
                                         Math.max(Double.parseDouble(context), LimitsHelper.PitchLimits.lowerLimit);
                                 tts.getMemberVoice(member).setPitchInAudioConfigBuilder(doubleValue);
@@ -123,7 +123,7 @@ public class VoiceCommands {
                             break;
                         default:
                             if (context.matches(RegexHelper.POS_DECIMAL_REGEX)) {
-                                final double doubleValue = Integer.parseInt(context) > 0 ?
+                                final double doubleValue = Double.parseDouble(context) > 0 ?
                                         Math.min(Double.parseDouble(context), LimitsHelper.SpeakingRateLimits.upperLimit) :
                                         Math.max(Double.parseDouble(context), LimitsHelper.SpeakingRateLimits.lowerLimit);
                                 tts.getMemberVoice(member).setSpeakingRateInAudioConfigBuilder(doubleValue);
